% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprev-package.R
\docType{package}
\name{rprev-package}
\alias{rprev}
\alias{rprev-package}
\title{rprev: Estimating Disease Prevalence from Registry Data}
\description{
Estimates disease prevalence for a given index date using existing registry data extended with Monte Carlo simulations following the method of Crouch et al (2014) \doi{ 10.1016/j.canep.2014.02.005}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stulacy/rprev-dev}
}

}
\author{
\strong{Maintainer}: Stuart Lacy \email{stuart.lacy@york.ac.uk}

Authors:
\itemize{
  \item Simon Crouch \email{simon.crouch@york.ac.uk}
  \item Stephanie Lax \email{stephanie.lax@york.ac.uk}
}

}
\keyword{internal}
