% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKmortality}
\alias{UKmortality}
\title{General population survival data.}
\format{
A data frame with 109575 rows and 3 columns:
\describe{
 \item{age}{age in days}
 \item{sex}{string, either 'M' or 'F'}
 \item{surv}{survival probability , estimated as the cumulative product of (1 - mortality rate)}
}
}
\usage{
UKmortality
}
\description{
A dataset containing daily population survival rates for individuals up to 100 years old,
from the UK population, derived from the 2009 mortality rates found at:
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/lifeexpectancies/datasets/nationallifetablesunitedkingdomreferencetables},
Adapted from public sector information licensed under the Open Government Licence v3.0.
Data were relabelled according to the mean year of the three-year birth window.
It is stored as a \code{data.table} for efficient access.
}
\keyword{datasets}
