% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_custom_models.R
\name{validate_incidence_model}
\alias{validate_incidence_model}
\title{Tests custom incidence models}
\usage{
validate_incidence_model(object, data, timeframe = 3652)
}
\arguments{
\item{object}{An incidence model to be tested}

\item{data}{Registry data in the form of a data frame. Ideally
will be the same source that will be used for the prevalence
estimation later on.}

\item{timeframe}{How long to generate incident cases for in days.
This is disease-specific, but the default of ten years
should work well for most diseases.}
}
\value{
The dummy incident population that has been generated to allow
for further diagnostics to be run.
}
\description{
Runs checks to assess whether a custom incidence model is
suitable for use in \code{prevalence}. Provides useful
diagnostic messages if any issues are encountered.
}
